Move struct declarations to types.go
