+++ pkg/certificate/providers/vault/client.go
@@ -4,7 +4,6 @@ import (
 	"time"
 
 	"github.com/hashicorp/vault/api"
-	"github.com/pkg/errors"
 
 	"github.com/open-service-mesh/osm/pkg/certificate"
 	"github.com/open-service-mesh/osm/pkg/constants"
@@ -19,24 +18,6 @@ const (
 	privateKeyField  = "private_key"
 	issuingCAField   = "issuing_ca"
 )
-
-// Client is a Hashi Vault client instance.
-type Client struct {
-	// How long will newly issued certificates be valid for
-	validity time.Duration
-
-	// The Certificate Authority root certificate to be used by this certificate manager
-	ca certificate.Certificater
-
-	// The channel announcing to the rest of the system when a certificate has changed
-	announcements <-chan interface{}
-
-	// Cache for all the certificates issued
-	cache map[certificate.CommonName]Certificate
-
-	// Hashicorp Vault Client
-	client *api.Client
-}
 
 // NewCertManager implements certificate.Manager and wraps a Hashi Vault with methods to allow easy certificate issuance.
 func NewCertManager(vaultAddr, token string, validity time.Duration) (*Client, error) {
@@ -73,7 +54,7 @@ func (c *Client) IssueCertificate(cn certificate.CommonName) (certificate.Certif
 	secret, err := c.client.Logical().Write(getIssueURL(), getIssuanceData(cn, c.validity))
 	if err != nil {
 		log.Error().Err(err).Msgf("Error issuing new certificate for CN=%s", cn)
-		return nil, errors.Wrap(err, "vault: issue:")
+		return nil, err
 	}
 
 	return newCert(cn, secret, time.Now().Add(c.validity)), nil
